Render the Twilight Struggle API docs in CI
Add bin/openapi, which merges the OpenAPI fragments into one document with
plain Ruby (no app boot, no DB), and an api_docs CI job that renders it to a
self-contained HTML page via Redocly and uploads it as an artifact.

Extract the schema merge out of the committee initializer into
lib/twilight_struggle_api_schema.rb so the middleware and bin/openapi share
one document builder. Shipped as bin/openapi rather than a rake task because
undercover scans .rake files as coverable code; an extensionless bin script
keeps the doc generator out of the coverage gate.
