+++ .github/workflows/ci.yml
@@ -193,3 +193,36 @@ jobs:
           git fetch --no-tags origin "$BASE_REF"
           bundle exec undercover --compare "origin/$BASE_REF"
 
+  api_docs:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v7
+        with:
+          persist-credentials: false
+
+      - name: Set up Ruby
+        uses: ruby/setup-ruby@003a5c4d8d6321bd302e38f6f0ec593f77f06600 # v1.319.0
+        with:
+          bundler-cache: true
+
+      - name: Set up Node
+        uses: actions/setup-node@v4
+        with:
+          node-version: 20
+
+      # bin/openapi merges the schema fragments with plain Ruby (no app boot, no
+      # DB), so this job stays a lightweight Ruby+Node runner like the other jobs.
+      - name: Generate the merged OpenAPI document
+        run: bin/openapi > openapi.json
+
+      - name: Render the API docs to a self-contained page
+        run: npx --yes @redocly/cli@1.34.5 build-docs openapi.json -o api-docs.html
+
+      - name: Upload the API docs
+        uses: actions/upload-artifact@v4
+        with:
+          name: twilight-struggle-api-docs
+          path: api-docs.html
+
