--- lib/tasks/twilight_struggle.rake
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-namespace :twilight_struggle do
-  desc "Print the merged Twilight Struggle OpenAPI document as JSON"
-  task openapi: :environment do
-    puts JSON.pretty_generate(TwilightStruggleApiSchema.document)
-  end
-end
