+++ lib/twilight_struggle_api_schema.rb
@@ -1,0 +1,40 @@
+# frozen_string_literal: true
+
+require "yaml"
+
+# Builds the Twilight Struggle API's OpenAPI document from the per-controller
+# fragments under config/api/twilight-struggle/v1/. Rails-free on purpose: the
+# `document` path (schema merge) is reused by the docs-generation rake task,
+# which must run without booting the app or a database.
+module TwilightStruggleApiSchema
+  DIR = File.expand_path("../config/api/twilight-struggle/v1", __dir__)
+
+  def self.driver
+    Committee::Drivers.load_from_data(document, parser_options: {strict_reference_validation: true})
+  end
+
+  def self.document
+    document = {
+      "openapi" => "3.0.3",
+      "info" => {"title" => "shrkbot Twilight Struggle API", "version" => "1.0"},
+      "paths" => {},
+      "components" => {"schemas" => {}}
+    }
+
+    Dir.glob(File.join(DIR, "*.yaml")).sort.each do |path|
+      fragment = YAML.safe_load_file(path)
+      merge_disjoint!(document["paths"], fragment.fetch("paths", {}), path)
+      merge_disjoint!(document["components"]["schemas"], fragment.dig("components", "schemas") || {}, path)
+    end
+
+    document
+  end
+
+  def self.merge_disjoint!(into, from, source)
+    from.each do |key, value|
+      raise "#{source}: #{key.inspect} is already defined by another schema fragment" if into.key?(key)
+
+      into[key] = value
+    end
+  end
+end
